requirements.txt
importlib
filelock
click==8.1.8
click-option-group==0.5.6
numpy==2.2.5
pandas==2.2.3
numexpr==2.11.0
joblib==1.4.2
scikit-learn==1.7.1
scipy==1.15.2
networkx==3.3
xgboost==3.0.2
Cython==3.0.12
h5py==3.12.1
contourpy==1.2.1
tensorflow==2.20.0
spliceai==1.3.1
biogl==2.3.0
biopython>=1.85
matplotlib
maturin==1.10.1
